+++ libsq/ast/segment.go
@@ -9,7 +9,14 @@ import (
 	"github.com/neilotoole/sq/libsq/ast/internal/slq"
 )
 
-var _ Node = (*SegmentNode)(nil)
+// VisitSegment implements slq.SLQVisitor.
+func (v *parseTreeVisitor) VisitSegment(ctx *slq.SegmentContext) any {
+	seg := newSegmentNode(v.ast, ctx)
+	v.ast.AddSegment(seg)
+	v.cur = seg
+
+	return v.VisitChildren(ctx)
+}
 
 func newSegmentNode(ast *AST, ctx *slq.SegmentContext) *SegmentNode {
 	seg := &SegmentNode{}
@@ -19,8 +26,11 @@ func newSegmentNode(ast *AST, ctx *slq.SegmentContext) *SegmentNode {
 	return seg
 }
 
+var _ Node = (*SegmentNode)(nil)
+
 // SegmentNode models a segment of a query (the elements separated by pipes).
-// For example, ".user | .uid, .username" is two segments (".user" and ".uid, .username").
+// For example, ".user | .uid, .username" is two segments: ".user",
+// and ".uid, .username".
 type SegmentNode struct {
 	bn baseNode
 }
@@ -121,6 +131,7 @@ func (s *SegmentNode) SegIndex() int {
 	return -1
 }
 
+// String returns a log/debug-friendly representation.
 func (s *SegmentNode) String() string {
 	if len(s.Children()) == 1 {
 		return fmt.Sprintf("segment[%d]: [1 element]", s.SegIndex())
@@ -129,6 +140,7 @@ func (s *SegmentNode) String() string {
 	return fmt.Sprintf("segment[%d]: [%d elements]", s.SegIndex(), len(s.Children()))
 }
 
+// Text implements ast.Node.
 func (s *SegmentNode) Text() string {
 	return s.bn.Context().GetText()
 }