+++ libsq/ast/selector.go
@@ -9,6 +9,36 @@ import (
 	"github.com/neilotoole/sq/libsq/ast/internal/slq"
 )
 
+// VisitSelectorElement implements slq.SLQVisitor.
+func (v *parseTreeVisitor) VisitSelectorElement(ctx *slq.SelectorElementContext) any {
+	node, err := newSelectorNode(v.cur, ctx.Selector())
+	if err != nil {
+		return err
+	}
+
+	if aliasCtx := ctx.Alias(); aliasCtx != nil {
+		node.alias = ctx.Alias().ID().GetText()
+	}
+
+	if err := v.cur.AddChild(node); err != nil {
+		return err
+	}
+
+	// No need to descend to the children, because we've already dealt
+	// with them in this function.
+	return nil
+}
+
+// VisitSelector implements slq.SLQVisitor.
+func (v *parseTreeVisitor) VisitSelector(ctx *slq.SelectorContext) any {
+	node, err := newSelectorNode(v.cur, ctx)
+	if err != nil {
+		return err
+	}
+
+	return v.cur.AddChild(node)
+}
+
 const (
 	msgNodeNoAddChild    = "%T cannot add children: failed to add %T"
 	msgNodeNoAddChildren = "%T cannot add children: failed to add %d children"
@@ -255,32 +285,25 @@ func (n *ColSelectorNode) String() string {
 	return str
 }
 
-var _ Node = (*Cmpr)(nil)
+var _ Node = (*CmprNode)(nil)
 
-// Cmpr models a comparison, such as ".age == 42".
-type Cmpr struct {
+// CmprNode models a comparison, such as ".age == 42".
+type CmprNode struct {
 	baseNode
 }
 
-func (c *Cmpr) String() string {
+// String returns a log/debug-friendly representation.
+func (c *CmprNode) String() string {
 	return nodeString(c)
 }
 
-func newCmpr(parent Node, ctx slq.ICmprContext) *Cmpr {
+func newCmprNode(parent Node, ctx slq.ICmprContext) *CmprNode {
 	leaf, _ := ctx.GetChild(0).(*antlr.TerminalNodeImpl) // FIXME: return an error
-	cmpr := &Cmpr{}
+	cmpr := &CmprNode{}
 	cmpr.ctx = leaf
+	cmpr.text = leaf.GetText()
 	cmpr.parent = parent
 	return cmpr
-}
-
-// HandleNode models a source handle such as "@sakila_sl3".
-type HandleNode struct {
-	baseNode
-}
-
-func (d *HandleNode) String() string {
-	return nodeString(d)
 }
 
 // extractSelVal extracts the value of the selector. The function takes